# docs/swagger.yaml
basePath: /api/v1
definitions:
  ginx.ResponseData:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  ginx.ResponseFail:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
...
  schema.DeptLabel:
    properties:
      children:
        items:
          $ref: '#/definitions/schema.DeptLabel'
        type: array
      id:
        type: integer
      label:
        type: string
    type: object
  schema.GetDeptTreeRoleSelectResp:
    properties:
      checkedKeys:
        items:
          type: integer
        type: array
      depts:
        items:
          $ref: '#/definitions/schema.DeptLabel'
        type: array
    type: object
  schema.GetMenuTreeSelectResp:
    properties:
      checkedKeys:
        items:
          type: integer
        type: array
      menus:
        items:
          $ref: '#/definitions/schema.MenuLabel'
        type: array
    type: object
...
  schema.MenuLabel:
    properties:
      children:
        items:
          $ref: '#/definitions/schema.MenuLabel'
        type: array
      id:
        type: integer
      label:
        type: string
    type: object
...
  schema.SysApi:
    properties:
      action:
        description: 请求类型
        type: string
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      deletedAt:
        description: 删除时间
        type: string
      handle:
        description: handle
        type: string
      id:
        description: 主键编码
        type: integer
      path:
        description: 地址
        type: string
      title:
        description: 标题
        type: string
      type:
        description: 接口类型
        type: string
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
    type: object
...
  schema.SysDept:
    properties:
      children:
        items:
          $ref: '#/definitions/schema.SysDept'
        type: array
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      dataScope:
        type: string
      deptId:
        description: 部门编码
        type: integer
      deptName:
        description: 部门名称
        type: string
      deptPath:
        type: string
      email:
        description: 邮箱
        type: string
      leader:
        description: 负责人
        type: string
      params:
        type: string
      parentId:
        description: 上级部门
        type: integer
      phone:
        description: 手机
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
    type: object
  schema.SysDeptDeleteReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  schema.SysDeptInsertReq:
    properties:
      createBy:
        description: 创建者
        type: integer
      deptName:
        description: 部门名称
        type: string
      deptPath:
        description: 路径
        type: string
      email:
        description: 邮箱
        type: string
      leader:
        description: 负责人
        type: string
      parentId:
        description: 上级部门
        type: integer
      phone:
        description: 手机
        type: string
      sort:
        description: 排序
        type: integer
      status:
        default: 2
        description: 状态
        type: integer
    required:
    - deptName
    - email
    - leader
    - parentId
    - phone
    - sort
    - status
    type: object
  schema.SysDeptUpdateReq:
    properties:
      deptId:
        description: 编码
        type: integer
      deptName:
        description: 部门名称
        type: string
      deptPath:
        description: 路径
        type: string
      email:
        description: 邮箱
        type: string
      leader:
        description: 负责人
        type: string
      parentId:
        description: 上级部门
        type: integer
      phone:
        description: 手机
        type: string
      sort:
        description: 排序
        type: integer
      status:
        default: 2
        description: 状态
        type: integer
      updateBy:
        description: 更新者
        type: integer
    required:
    - deptName
    - email
    - leader
    - parentId
    - phone
    - sort
    - status
    type: object
  schema.SysMenu:
    properties:
      action:
        type: string
      apis:
        items:
          type: integer
        type: array
      breadcrumb:
        type: integer
      children:
        items:
          $ref: '#/definitions/schema.SysMenu'
        type: array
      component:
        type: string
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      dataScope:
        type: string
      icon:
        type: string
      is_select:
        type: boolean
      isFrame:
        type: integer
      menuId:
        type: integer
      menuName:
        type: string
      menuType:
        type: string
      noCache:
        type: integer
      params:
        type: string
      parentId:
        type: integer
      path:
        type: string
      paths:
        type: string
      permission:
        type: string
      roleId:
        type: integer
      sort:
        type: integer
      sysApi:
        items:
          $ref: '#/definitions/schema.SysApi'
        type: array
      title:
        type: string
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
      visible:
        type: integer
    type: object
  schema.SysMenuDeleteReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  schema.SysMenuInsertReq:
    properties:
      action:
        description: 请求方式
        type: string
      breadcrumb:
        description: 是否面包屑
        type: integer
      component:
        description: 组件
        type: string
      createBy:
        type: integer
      icon:
        description: 图标
        type: string
      isFrame:
        description: 是否frame
        type: integer
      menuName:
        description: 菜单name
        type: string
      menuType:
        description: 菜单类型
        type: string
      noCache:
        description: 是否缓存
        type: integer
      parentId:
        description: 上级菜单
        type: integer
      path:
        description: 路径
        type: string
      paths:
        description: id路径
        type: string
      permission:
        description: 权限编码
        type: string
      sort:
        description: 排序
        type: integer
      title:
        description: 显示名称
        type: string
      visible:
        description: 是否显示
        type: integer
    type: object
  schema.SysMenuUpdateReq:
    properties:
      action:
        description: 请求方式
        type: string
      apis:
        items:
          type: integer
        type: array
      breadcrumb:
        description: 是否面包屑
        type: integer
      component:
        description: 组件
        type: string
      icon:
        description: 图标
        type: string
      id:
        description: 编码
        type: integer
      isFrame:
        description: 是否frame
        type: string
      menuName:
        description: 菜单name
        type: string
      menuType:
        description: 菜单类型
        type: string
      noCache:
        description: 是否缓存
        type: integer
      parentId:
        description: 上级菜单
        type: integer
      path:
        description: 路径
        type: string
      paths:
        description: id路径
        type: string
      permission:
        description: 权限编码
        type: string
      sort:
        description: 排序
        type: integer
      title:
        description: 显示名称
        type: string
      updateBy:
        type: integer
      visible:
        description: 是否显示
        type: integer
    type: object
...
host: 127.0.0.1:8000
info:
  contact:
    email: conjurer888888@gmail.com
    name: conjurer
    url: https:/github.com/dot123
  description: This is a game management background. you can use the api key `ApiKeyAuth`
    to test the authorization filters.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com
  title: gin-gorm-admin API
  version: "1.0"
paths:
  /dept:
    delete:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysDeptDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 删除部门
      tags:
      - 部门
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: query
        name: deptID
        type: integer
      - description: 部门名称
        in: query
        name: deptName
        type: string
      - description: 路径
        in: query
        name: deptPath
        type: string
      - description: 邮箱
        in: query
        name: email
        type: string
      - description: 负责人
        in: query
        name: leader
        type: string
      - description: 上级部门
        in: query
        name: parentID
        type: integer
      - description: 手机
        in: query
        name: phone
        type: string
      - description: 排序
        in: query
        name: sort
        type: integer
      - description: 状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.DeptLabel'
                  type: array
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 分页部门列表数据
      tags:
      - 部门
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysDeptInsertReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 添加部门
      tags:
      - 部门
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysDeptUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 修改部门
      tags:
      - 部门
  /dept/{deptId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: deptId
        in: path
        name: deptId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.SysDept'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 获取部门数据
      tags:
      - 部门
  /deptTree:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: query
        name: deptID
        type: integer
      - description: 部门名称
        in: query
        name: deptName
        type: string
      - description: 路径
        in: query
        name: deptPath
        type: string
      - description: 邮箱
        in: query
        name: email
        type: string
      - description: 负责人
        in: query
        name: leader
        type: string
      - description: 上级部门
        in: query
        name: parentID
        type: integer
      - description: 手机
        in: query
        name: phone
        type: string
      - description: 排序
        in: query
        name: sort
        type: integer
      - description: 状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.DeptLabel'
                  type: array
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 左侧部门树
      tags:
      - 部门
...
  /menu:
    delete:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysMenuDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 删除菜单
      tags:
      - 菜单
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: pageNum
        type: integer
      - in: query
        maximum: 100
        name: pageSize
        type: integer
      - description: 菜单名称
        in: query
        name: title
        type: string
      - description: 显示状态
        in: query
        name: visible
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.SysMenu'
                  type: array
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: Menu列表数据
      tags:
      - 菜单
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysMenuInsertReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 创建菜单
      tags:
      - 菜单
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysMenuUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 修改菜单
      tags:
      - 菜单
  /menu/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.SysMenu'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: Menu详情数据
      tags:
      - 菜单
  /menurole:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.SysMenu'
                  type: array
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 根据登录角色名称获取菜单列表数据（左菜单使用）
      tags:
      - 菜单
...
  /roleDeptTreeselect/{roleId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: roleId
        in: path
        name: roleId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetDeptTreeRoleSelectResp'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 部门树形
      tags:
      - 部门
  /roleMenuTreeselect/{roleId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: roleId
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetMenuTreeSelectResp'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 根据角色ID查询菜单下拉树结构
      tags:
      - 菜单
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"